+++ kotori/frontend/server.py
@@ -7,10 +7,13 @@ from twisted.web.wsgi import WSGIResource
 from pyramid.paster import get_app
 
 
-def boot_frontend(http_port, websocket_uri, debug=False):
+def boot_frontend(config, debug=False):
     """
     Boot a Pyramid WSGI application as Twisted component
     """
+
+    http_port = int(config.get('config-web', 'http_port'))
+    websocket_uri = unicode(config.get('wamp', 'listen'))
 
     # https://stackoverflow.com/questions/13122519/serving-pyramid-application-using-twistd/13138610#13138610
     config = resource_filename('kotori.frontend', 'development.ini')