+++ CHANGES.rst
@@ -4,6 +4,10 @@ Kotori CHANGES
 
 develop
 =======
+- Improve configuration file, logging, debugging and documentation
+- Improve internal settings handling and application bootstrapping
+- Integrate subsystems from vendor Hiveeyes/BERadio into Kotori core
+- Document how to run on PyPy
 
 
 2016-03-27 0.6.0